🖋 Update

+++ builder.js
@@ -14,10 +14,14 @@ const rimraf = require('rimraf');
 
 if(!fs.existsSync('dist')){
     fs.mkdirSync('dist');
+    getFiles('pages');
 }
 else{
-    fs.rmdirSync('dist', { recursive: true });
-    fs.mkdirSync('dist');
+    fs.rm('dist', {recursive: true}, ()=>{
+        console.log('Clean !');
+        fs.mkdirSync('dist');
+        getFiles('pages');
+    });
 }
 if(fs.existsSync('pages/demo')){
     rimraf('pages/demo', (err) => {
@@ -38,40 +42,42 @@ function getFiles (dir, files_){
     files_ = files_ || [];
     var files = fs.readdirSync(dir);
     for (var i in files){
-        //PAGES
-        var config = require(`./pages/${files[i]}/${files[i]}.json`);
-        const $ = cheerio.load(fs.readFileSync('index.html'));
-        
-        //TITLES
-        $('title').text(`Lyna | ${config.name}`);
-        $('meta[name="title"]').attr('content', `Lyna | ${config.name}`);
-        $('meta[name="twitter:title"]').attr('content', `Lyna | ${config.name}`);
-        $('meta[property="og:title"]').attr('content', `Lyna | ${config.name}`);
-        
-        //IMAGES
-        $('meta[name="twitter:image"]').attr('content', `https://lyna.netlify.app/${config.logo}`);
-        $('meta[property="og:image"]').attr('content', `https://lyna.netlify.app/${config.logo}`);
-
-        //DESCRIPTION
-        if(config.description == null){config.description = "Create a page for provide your social links and other."}
-        $('meta[name="description"]').attr('content', `${config.description}`);
-        $('meta[property="og:description"]').attr('content', `${config.description}`);
-        $('meta[name="twitter:description"]').attr('content', `${config.description}`);
-        
-        fs.writeFileSync(`dist/${files[i]}.html`, $.html());
-
-        sitemap("urlset").append(`
-            <url>
-                <loc>https://lyna.ga/${files[i]}</loc>
-                <lastmod>${today}</lastmod>
-                <changefreq>hourly</changefreq>
-            </url>
-        `);
-
-        fs.writeFileSync(`./dist/sitemap.xml`, sitemap.html());
-        fs.writeFileSync(`sitemap.xml`, sitemap.html());
-        fs.writeFileSync(`sitemap2021.xml`, sitemap.html());
+        if(fs.existsSync(`./pages/${files[i]}/${files[i]}.json`)){
+            //PAGES
+            var config = require(`./pages/${files[i]}/${files[i]}.json`);
+            const $ = cheerio.load(fs.readFileSync('index.html'));
+            
+            //TITLES
+            $('title').text(`Lyna | ${config.name}`);
+            $('meta[name="title"]').attr('content', `Lyna | ${config.name}`);
+            $('meta[name="twitter:title"]').attr('content', `Lyna | ${config.name}`);
+            $('meta[property="og:title"]').attr('content', `Lyna | ${config.name}`);
+            
+            //IMAGES
+            $('meta[name="twitter:image"]').attr('content', `https://lyna.netlify.app/${config.logo}`);
+            $('meta[property="og:image"]').attr('content', `https://lyna.netlify.app/${config.logo}`);
+            $('head').append(`<link rel="shortcut icon" href="https://lyna.ga/${config.logo}" type="image/x-icon">`);
+    
+            //DESCRIPTION
+            if(config.description == null){config.description = "Create a page for provide your social links and other."}
+            $('meta[name="description"]').attr('content', `${config.description}`);
+            $('meta[property="og:description"]').attr('content', `${config.description}`);
+            $('meta[name="twitter:description"]').attr('content', `${config.description}`);
+            
+            fs.writeFileSync(`dist/${files[i]}.html`, $.html());
+    
+            sitemap("urlset").append(`
+                <url>
+                    <loc>https://lyna.ga/${files[i]}</loc>
+                    <lastmod>${today}</lastmod>
+                    <changefreq>hourly</changefreq>
+                </url>
+            `);
+    
+            fs.writeFileSync(`./dist/sitemap.xml`, sitemap.html());
+            fs.writeFileSync(`sitemap.xml`, sitemap.html());
+            fs.writeFileSync(`sitemap2021.xml`, sitemap.html());
+        }
     }
     return files_;
 }
-getFiles('pages');
